did some file and model renaming because it was bothering me.
